server/requirements.txt
fastapi>=0.95.2
uvicorn>=0.22.0
pydantic>=1.10.13
python-dotenv>=1.0.1
requests>=2.31.0
playwright>=1.42.0
python-multipart>=0.0.9
websockets>=12.0
langchain-google-genai>=0.0.11
langchain-core>=0.1.27
typing-extensions>=4.5.0
langchain>=0.1.0
langchain-community>=0.0.16
pylint>=3.0.0
google-generativeai>=0.3.2
langgraph>=0.1.10
mcp-python>=0.1.0
pytest
pytest-asyncio
httpx
youtube_search>=2.1.2
python-dateutil>=2.8.2